ai/requirements.txt
# AI Microservice Dependencies

# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML Models
torch==2.1.1
transformers==4.36.0
sentence-transformers==2.2.2
openai-whisper==20231117
accelerate==0.25.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# NLP and Entity Extraction
spacy==3.7.2
spacy-transformers==1.3.4

# Vector Database
weaviate-client==3.25.3

# HTTP clients
httpx==0.25.2
requests==2.31.0

# Data processing
numpy==1.25.2
pandas==2.1.4
pillow==10.1.0

# Ollama client for Llama
ollama==0.1.7

# YouTube integration
youtube-transcript-api==0.6.1
google-api-python-client==2.110.0

# Date/time parsing
python-dateutil==2.8.2

# Email validation
email-validator==2.1.0.post1

# Phone number parsing
phonenumbers==8.13.26

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
redis==5.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
